+++ R/estimates_dynamics.R
@@ -12,11 +12,7 @@
 #' @param condition name of column in dataframe data that specifies the experimental condition
 #' @param time column in "data" that contains the time point identifiers
 #' @param metabolite column of "data" that contains the metabolite names or IDs
-#' @param fits list of model fits for which estimates should be extracted
-#' @param iter how many iterations were used to fit the dynamics model
-#' @param warmup how many warm-up iterations were used to fit the dynamics model
-#' @param chains how many chains were used to fit the dynamics model
-#' @param samples how many posterior samples should be drawn (p.e. for check of clustering precision)
+#' @param fit model fit for which estimates should be extracted
 #'
 #' @seealso Fit the dynamic model [fit_dynamics_model()].
 #' Diagnostics of the dynamic model [diagnostics_dynamics()]
@@ -27,12 +23,13 @@
 #' @import SummarizedExperiment
 #' @importFrom S4Vectors metadata
 #'
-#' @return a list of dataframes (one per experimental condition) that contains
-#' the estimates at the timepoints and samples from the posterior
-#' (number as specified in samples), delta_mu specifies the difference between
-#' time point specified in column "time.ID" and subsequent time point (delta_mu
-#' in row time.ID=1: mu(time point 2)- mu(time point 1)) if number of time points
-#' in dataset is >1
+#' @return a list of dataframes (one per parameters mu, sigma, lambda, delta_mu and euclidean distance) that contains
+#' the estimates:
+#' mu: is the estimated mean metabolite abundance
+#' sigma: the estimated standard deviation of metabolite abundance
+#' lambda: pooled sigma per condition
+#' delta_mu: differences of mu between time points
+#' euclidean_distance: estimated euclidean distance of time vectors of one metabolite between conditions
 #' If data is a summarizedExperiment object the estimates are stored in
 #' metadata(data) under "estimates_dynamics"
 #' @export
@@ -40,24 +37,22 @@
 #' @examples
 #' data("longitudinalMetabolomics")
 #' data <- longitudinalMetabolomics[, longitudinalMetabolomics$condition == "A" &
-#'   longitudinalMetabolomics$metabolite == "ATP"]
+#'                                    longitudinalMetabolomics$metabolite == "ATP"]
 #' data <- fit_dynamics_model(
 #'   data = data,
 #'   scaled_measurement = "m_scaled", assay = "scaled_log",
 #'   max_treedepth = 14, adapt_delta = 0.95, iter = 2000, cores = 1, chains = 1
 #' )
-#' data <- estimates_dynamics(
-#'   data = data, iter = 2000,
-#'   chains = 1, condition = "condition"
+#'data <- estimates_dynamics(
+#'   data = data
 #' )
-#  S4Vectors::metadata(data)[["estimates_dynamics"]]
-#'
+#' S4Vectors::metadata(data)[["estimates_dynamics"]]
 estimates_dynamics <- function(data, assay = "scaled_log",
                                kegg = "KEGG", condition = "condition", time = "time",
                                metabolite = "metabolite",
-                               fits = metadata(data)[["dynamic_fits"]],
-                               iter = 2000, warmup = iter / 4, chains = 4,
-                               samples = 1) {
+                               fit = metadata(data)[["dynamic_fit"]]) {
+  
+  
   # Input checks
   if (!is.data.frame(data) & !inherits(data, "SummarizedExperiment")) {
     stop("'data' must be a dataframe or colData of a SummarizedExperiment object")
@@ -73,9 +68,9 @@ estimates_dynamics <- function(data, assay = "scaled_log",
       cols = seq_len(t), names_to = "time",
       values_to = "scaled_measurement"
     )
-    fits <- metadata(data)[["dynamic_fits"]]
+    fit <- metadata(data)[["dynamic_fit"]]
   }
-
+  
   # convert potential tibbles into data frame
   if (is(data, "tbl")) {
     data <- as.data.frame(data)
@@ -84,8 +79,8 @@ estimates_dynamics <- function(data, assay = "scaled_log",
     data_df <- data
   }
   # check if all elements of fits are stanfit objects
-  if (!all(vapply(fits, function(x) inherits(x, "stanfit"), logical(1)))) {
-    stop("'fits' must be a list of stanfit objects")
+  if (!inherits(fit, "stanfit")) {
+    stop("'fit' must be a stanfit objects")
   }
   if (!all(vapply(list(time, condition, kegg), is.character, logical(1)))) {
     stop("'time', 'kegg' and 'condition' must be a character vector specifying a column name of data")
@@ -93,113 +88,105 @@ estimates_dynamics <- function(data, assay = "scaled_log",
   if (!all(c(time, kegg, condition) %in% colnames(data_df))) {
     stop("'data' must contain columns named 'condition', 'kegg' and 'time'")
   }
-  if (!(all(c(warmup, iter, chains, samples) > 0 & c(warmup, iter, chains, samples) %% 1 == 0))) {
-    stop("'iter', 'warmup', 'chains', and 'samples' must be positive integers")
-  }
 
+  # get number of metabolites, time points and conditions
   M <- length(unique(data_df[[metabolite]]))
   t <- length(unique(data_df[[time]]))
-  # Unique conditions
-  conditions <- unique(data[[condition]])
+  C <- length(unique(data[[condition]]))
 
-  # Apply helper function to all conditions
-  dynamics <- lapply(conditions, function(cond) {
-    fit <- fits[[cond]]
-    pS <- as.data.frame(rstan::summary(fit)$summary)
-    mu_posterior <- as.matrix(fit)
 
-    # Precompute random draws
-    random_draws <- floor(runif(samples, min = 1, max = (iter - warmup) * chains))
+  estimates_data <- data.frame(
+    metabolite = rep(rep(unique(data_df$metabolite),each=C),each=t),
+    time = rep(rep(unique(data_df$time),each=C),M),
+    condition = rep(rep(unique(data_df$condition),t),M)
+  )
+  
+  # extract for mu
+  mu <- rstan::summary(fit,pars="mu")$summary
+  mu <- cbind(estimates_data, parameter = "mu",mu[,c("mean","2.5%","97.5%")])
 
-    # Generate all parameter combinations
-    mu_indices <- expand.grid(metabolite = seq_len(M), timepoint = seq_len(t))
-    mu_indices_names <- paste0(
-      "mu[", mu_indices$metabolite,
-      ",", mu_indices$timepoint, "]"
+  # extract for sigma
+  sigma <- rstan::summary(fit,pars="sigma")$summary
+  sigma <- cbind(estimates_data, parameter = "sigma", sigma[,c("mean","2.5%","97.5%")])
+  
+  # extract for lambda
+  ## lambda only one per condition -> adapt estimates_data
+  lambda_data <- data.frame(
+    metabolite = rep(unique(data_df$metabolite),each=C),
+    condition = rep(unique(data_df$condition),M)
+  )
+
+  lambda <- rstan::summary(fit,pars="lambda")$summary
+  lambda <- cbind(lambda_data, parameter = "lambda", lambda[,c("mean","2.5%","97.5%")])
+  
+  # extract euclidean distances
+  ## get possible dose combinations
+  if(C>1&t>1){
+  combinations <- t(combn(unique(data_df$condition), 2))
+
+  distances_data <- data.frame(
+    metabolite = rep(unique(data_df$metabolite),each=nrow(combinations)),
+    condition_1 = combinations[,1],
+    condition_2 = combinations[,2]
+  )
+  
+  distances <- as.data.frame(rstan::summary(fit,pars="euclidean_distance")$summary)
+  distances <- distances%>%na.omit() # posterior contains estiamtes that do not fullfull condition_1<conditon_2 due to stan technicalities
+  distances <- cbind(distances_data,parameter="euclidean_distance",distances[,c("mean","2.5%","97.5%")])
+  }
+  if(t==1|C==1){
+    distances <- "only possible for more than one time point and more than one condition"
+  }
+  
+  # extract for delta_mu
+  if (t > 1) {
+    combinations <- t(combn(unique(data_df$time), 2))
+
+    delta_mu_data <- data.frame(
+      metabolite = rep(rep(unique(data_df$metabolite),each=nrow(combinations)),each=C),
+      condition = rep(rep(unique(data_df$condition),each=nrow(combinations)),M),
+      timepoint_1 = combinations[,1],
+      timepoint_2 = combinations[,2]
     )
+  
+  delta_mu <- as.data.frame(rstan::summary(fit,pars="delta_mu")$summary)
+  delta_mu <- delta_mu%>%na.omit()
+  delta_mu <- cbind(delta_mu_data,parameter="delta_mu",delta_mu[,c("mean","2.5%","97.5%")])
+  
+  }
+  if(t==1){
+    delta_mu <- "only possible for more than one time point"
+  }
+    
+  # combine results in one list
+  result <- list(mu = mu, 
+                 sigma = sigma , 
+                 lambda = lambda, 
+                 delta_mu= delta_mu, 
+                 euclidean_distance=distances)
 
-    sigma_indices <- expand.grid(metabolite = seq_len(M), timepoint = seq_len(t))
-    sigma_indices_names <- paste0(
-      "sigma[", sigma_indices$metabolite,
-      ",", sigma_indices$timepoint, "]"
-    )
-    lambda_indices_names <- paste0("lambda[", seq_len(M), "]")
-    delta_mu_indices <- expand.grid(metabolite = seq_len(M), timepoint = seq_len(t - 1))
-    delta_mu_indices_names <- paste0(
-      "delta_mu[", delta_mu_indices$metabolite,
-      ",", delta_mu_indices$timepoint, "]"
-    )
+  # Map metabolite IDs to KEGG and names
+  metabolites <- unique(data_df[c(metabolite, kegg)])
 
-    # Extract and format results
-    mu_data <- data.frame(
-      metabolite.ID = rep(seq_len(M), each = t),
-      time.ID = rep(seq_len(t), M),
-      condition = cond,
-      mu_mean = pS[mu_indices_names, "mean"],
-      mu_lower = pS[mu_indices_names, "2.5%"],
-      mu_higher = pS[mu_indices_names, "97.5%"]
-    )
-
-    sigma_data <- data.frame(
-      sigma_mean = pS[sigma_indices_names, "mean"],
-      sigma_lower = pS[sigma_indices_names, "2.5%"],
-      sigma_higher = pS[sigma_indices_names, "97.5%"]
-    )
-
-    lambda_data <- data.frame(
-      metabolite.ID = seq_len(M),
-      lambda_mean = pS[lambda_indices_names, "mean"],
-      lambda_lower = pS[lambda_indices_names, "2.5%"],
-      lambda_higher = pS[lambda_indices_names, "97.5%"]
-    )
-
-    if (t > 1) {
-      delta_mu_data <- data.frame(
-        metabolite.ID = rep(seq_len(M), t - 1),
-        time.ID = rep(seq_len(t - 1), each = M), # one less delta_t than timepoints
-        delta_mu_mean = pS[delta_mu_indices_names, "mean"],
-        delta_mu_lower = pS[delta_mu_indices_names, "2.5%"],
-        delta_mu_higher = pS[delta_mu_indices_names, "97.5%"]
-      )
+  # # Join metadata: for all list elements
+  result <- lapply(result,function(x){
+    if(is.data.frame(x)&metabolite%in%colnames(x)){
+      x <- left_join(x,metabolites,by=metabolite)
+      x <- x %>% relocate (!!kegg, .after=!!metabolite)
     }
-
-    # Combine results into a single data frame
-    result <- cbind(mu_data, sigma_data)
-
-    # Add lambda and delta_mu values
-    result <- left_join(result, lambda_data, by = "metabolite.ID")
-    if (t > 1) {
-      result <- left_join(result, delta_mu_data, by = c("metabolite.ID", "time.ID"))
-    }
-    # Add samples
-    sample_data <- vapply(random_draws, function(draw) {
-      vapply(mu_indices_names, function(name) mu_posterior[draw, name], numeric(1))
-    }, numeric(length(mu_indices_names)))
-    colnames(sample_data) <- paste0("mu_sample_", seq_len(samples))
-    rownames(sample_data) <- NULL
-    result <- cbind(result, sample_data)
-
-    # Map metabolite IDs to KEGG and names
-    metabolites <- unique(data_df[c("metabolite", kegg)])
-    metabolites$metabolite.ID <- as.numeric(as.factor(metabolites$metabolite))
-
-    # Join metadata
-    result <- left_join(result, metabolites, by = "metabolite.ID")
-    result <- result %>% relocate("metabolite", .after = "metabolite.ID")
-    result <- result %>% relocate(kegg, .after = "metabolite")
-    result <- unique(result)
-
-    return(result)
+    return(x)
   })
+  # result <- left_join(result, metabolites, by = "metabolite")
+  # result <- result %>% relocate("metabolite", .after = "metabolite")
+  # result <- result %>% relocate(kegg, .after = "metabolite")
+  # result <- unique(result)
 
-  names(dynamics) <- conditions
-
-  # if input is a SummarizedExperiment object, store the fits in the metadata
+  # if input is a SummarizedExperiment object, store estimates in the metadata
   if (is(data, "SummarizedExperiment")) {
-    metadata(data)[["estimates_dynamics"]] <- dynamics
+    metadata(data)[["estimates_dynamics"]] <- result
     return(data)
   } else {
     # otherwise, return the list of fits
-    return(dynamics)
+    return(result)
   }
 }